refactor: move Cypress test helpers for list layouts to ui_test_helpers and update references
